Split de/serialize functions into common and content rules; Fix error handling for parseContentHtml
