+++ src/cnxml/content/serialize.js
@@ -34,43 +34,6 @@ const BLOCK = {
 
 
 /**
- * Serialize a mark.
- *
- * @see MARK_TAGS
- */
-const MARK = {
-    serialize(obj, children) {
-        const Mark = MARK_TAGS[obj.type]
-        if (!Mark) return
-
-        if (Mark instanceof Function) return Mark(obj, children)
-
-        return <Mark {...obj.data.toJS()}>
-            {children}
-        </Mark>
-    },
-}
-
-
-/**
- * Serialize text.
- *
- * `slate-html-serializer` replaces newlines in text with `<br>` HTML elements,
- * which are not supported in CNXML. Instead we just emit newlines, as they
- * don't pose any problems in CNXML.
- *
- * As an additional benefit, this makes it simpler to write serializers for
- * nodes, such as `code`, in which newlines have special meaning and must
- * be preserved.
- */
-const TEXT = {
-    serialize(obj, children) {
-        if (obj.object === 'string') return [children]
-    }
-}
-
-
-/**
  * Tags which can occur in block content.
  *
  * Keys are Slate node types, values are _transformer functions_. Transformer
@@ -96,7 +59,6 @@ const BLOCK_TAGS = {
     list_item: 'item',
     media: 'media',
     ol_list: list,
-    paragraph: 'para',
     quotation: 'quote',
     section: 'section',
     title: 'title',
@@ -105,30 +67,19 @@ const BLOCK_TAGS = {
 
 
 /**
- * Serializer for CNXML emphasis tag.
+ * Serializer for code.
  */
-const emphasis = type => function(obj, children) {
-    return <emphasis effect={type}>{children}</emphasis>
-}
+function code(obj, children) {
+    let attrs = {}
 
+    if (obj.object === 'block') {
+        attrs.display = 'block'
+        attrs.id = obj.key
+    }
 
-/**
- * Mark serialisation works similarly to block serialisation, except the default
- * function (the one used when value is a string) doesn't produce the
- * `id` attribute.
- *
- * @see MARK
- * @see BLOCK_TAGS
- */
-const MARK_TAGS = {
-    emphasis: emphasis('italics'),
-    link: 'link',
-    strong: emphasis('bold'),
-    subscript: 'sub',
-    superscript: 'sup',
-    term: term,
-    underline: emphasis('underline'),
-    xref: xref,
+    return <code {...attrs}>
+        {children}
+    </code>
 }
 
 
@@ -162,62 +113,6 @@ function list(obj, children) {
 }
 
 
-/**
- * Serializer for cross-references.
- */
-function xref(obj, children) {
-    let attrs = {
-        'target-id': obj.data.get('target'),
-    }
-
-    const cmlnleCase = obj.data.get('case')
-
-    if (cmlnleCase) {
-        attrs['cmlnleCase'] = cmlnleCase
-    }
-
-    return <link {...attrs}>
-        {children}
-    </link>
-}
-
-
-/**
- * Serializer for code.
- */
-function code(obj, children) {
-    let attrs = {}
-
-    if (obj.object === 'block') {
-        attrs.display = 'block'
-        attrs.id = obj.key
-    }
-
-    return <code {...attrs}>
-        {children}
-    </code>
-}
-
-
-/**
- * Serializer for terms.
- */
-function term(obj, children) {
-    let attrs = {}
-
-    const reference = obj.data.get('reference')
-    if (reference && reference !== obj.text) {
-        attrs['cmlnleReference'] = reference
-    }
-
-    return <term {...attrs}>
-        {children}
-    </term>
-}
-
-
 export default [
     BLOCK,
-    MARK,
-    TEXT,
 ]
